replication: handle multiple PARTIAL_UPDATE_ROWS_EVENT diffs
The logic for PARTIAL_UPDATE_ROWS_EVENT handling introduced in
April 2023 by commit 3d2be208 is incorrect because it assumes a
partial JSON update can only ever patch one location in the JSON
column value. This is not the case, so the correct representation
of a partial update value is a list of multiple diffs and the
parsing logic needs to loop until all diffs have been decoded.

I have not given much thought to migration here. Arguably it may
be helpful to add a special case so that a single-location diff
still results in a `*JsonDiff` value, but on the other hand that
is of limited usefulness because any consumers of partial JSON
diffs need to handle the multiple-patch case to be correct anyway
and if they're going to add proper handling then that special case
just forces the recipient to duplicate the value processing logic.